+++ config/projects/chef-server.rb
@@ -32,14 +32,13 @@ override :berkshelf2, version: "2.0.18"
 override :rabbitmq, version: "3.3.4"
 override :erlang, version: "R16B03-1"
 override :ruby, version: "2.1.4"
-override :'chef-gem', version: "12.0.3"
 override :'server-jre', version: "7u25"
 
 # creates required build directories
 dependency "preparation"
 
 # global
-dependency "chef-gem" # for embedded chef-client -z runs
+dependency "chef" # for embedded chef-client -z runs
 dependency "private-chef-scripts" # assorted scripts used by installed instance
 dependency "private-chef-ctl" # additional project-specific private-chef-ctl subcommands
 dependency "ctl-man" # install man page